+++ src/app/(client)/meal/post/edit/_components/add-meal-card-drawer.tsx
@@ -14,9 +14,9 @@ import { generateCaloriesAnalysisByText } from '@/apis/gemini.api';
 import { parseGeminiResponse } from '@/lib/gemini';
 import { AiResponseDTO } from '@/types/DTO/ai_analysis.dto';
 import { formatNumberWithComma } from '@/utils/format.util';
-import { MAX_MENU_NAME_LENGTH, MAX_NUMERIC_LENGTH } from '../constants/meal-edit.constant';
-import { parseNumber } from '../utils/meal-edit.util';
-import { ERROR_MESSAGES } from '../constants/error-message.constant';
+import { MAX_MENU_NAME_LENGTH, MAX_NUMERIC_LENGTH } from '../_constants/meal-edit.constant';
+import { parseNumber } from '../_utils/meal-edit.util';
+import { ERROR_MESSAGES } from '../_constants/error-message.constant';
 
 type FoodFormValues = {
   menuName: string;
