+++ src/app/(client)/meal/post/edit/_components/edit-result-section.tsx
@@ -5,7 +5,7 @@ import { MealCategory } from '@/types/meal-category.type';
 import { createMealWithDetails } from '@/apis/meal.api';
 import { uploadImage } from '@/apis/storage.api';
 import { formatTimestamp } from '@/utils/format.util';
-import { urlToFile } from '../../../utils/file.util';
+import { urlToFile } from '../../../_utils/file.util';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
 import { FormProvider, useFieldArray, useForm } from 'react-hook-form';
@@ -22,17 +22,17 @@ import MacronutrientBox from '@/components/commons/macronutrient-box';
 import MacronutrientPieChart from '@/components/commons/macronutrient-pie-chart';
 import { MacronutrientEnum } from '@/types/nutrition.type';
 import AddMealCardDrawer from './add-meal-card-drawer';
-import { MAX_MEMO_LENGTH } from '../constants/meal-edit.constant';
+import { MAX_MEMO_LENGTH } from '../_constants/meal-edit.constant';
 import { MEAL_CATEGORY } from '../../../_constants/category.constant';
 import MemoBox from '../../../[id]/_components/memo-box';
 import Responsive from '@/components/commons/responsive';
 import TimePickerPc from '../../../_components/time-picker-pc';
 import AddMealCardPc from './add-meal-card-pc';
 import Modal from '@/components/commons/modal';
-import { ERROR_MESSAGES } from '../constants/error-message.constant';
+import { ERROR_MESSAGES } from '../_constants/error-message.constant';
 import { CreateMealDTO } from '@/types/DTO/meal.dto';
 import { deleteAnalysisData } from '@/apis/analysis-request.api';
-import { ErrorMessage, handleError } from '../../../utils/error.util';
+import { ErrorMessage, handleError } from '../../../_utils/error.util';
 import MealCalendarPc from '../../../_components/meal-calendar-pc';
 import MealCalendarDrawer from '../../../_components/meal-calendar-drawer';
 import TimePicker, { TimeFields } from '../../../_components/time-picker';
